Move the K2Nodes to their own folder
